% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish2webs-package.R
\docType{package}
\name{fish2webs-package}
\alias{fish2webs}
\alias{fish2webs-package}
\title{fish2webs package}
\description{
Tools for transforming fish survey measurements into size-structured food webs and associated summary outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Fellows-of-the-Fish-Food-Webs-F3W/fish2webs}
  \item Report bugs at \url{https://github.com/Fellows-of-the-Fish-Food-Webs-F3W/fish2webs/issues}
}

}
\author{
\strong{Maintainer}: Willem Bonnaffe \email{w.bonnaffe@gmail.com}

}
\keyword{internal}
